extract calculateAndValidateDependentNodes
